Refactored crawler index, added crawl-all
