+++ package.json
@@ -4,7 +4,7 @@
     "email": "serg.s.r@gmail.com"
   },
   "scripts": {
-    "crawl": "bun ./crawler",
+    "crawl": "bun ./crawler/index.ts",
     "mrproper": "rimraf ./store/data.db*",
     "lint": "bun eslint . --fix",
     "serve": "bun ./server"
